Move client builders to major specific sub packages
