+++ http-utils/src/main/java/ai/vespa/util/http/hc4/VespaHttpClientBuilder.java
@@ -1,5 +1,5 @@
 // Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package ai.vespa.util.http;
+package ai.vespa.util.http.hc4;
 
 import com.yahoo.security.tls.MixedMode;
 import com.yahoo.security.tls.TlsContext;
